Updated from upstream.
git-svn-id: https://svn.osgeo.org/gdal/trunk@7305 f0d54148-0727-0410-94bb-9a71ac55c965

+++ frmts/gtiff/libtiff/tif_color.c
@@ -121,14 +121,14 @@ TIFFCIELabToRGBInit(TIFFCIELabToRGB* cielab,
 		    TIFFDisplay *display, float *refWhite)
 {
 	int i;
-	float gamma;
+	double gamma;
 
 	cielab->range = CIELABTORGB_TABLE_RANGE;
 
 	_TIFFmemcpy(&cielab->display, display, sizeof(TIFFDisplay));
 
 	/* Red */
-	gamma = 1.0F / cielab->display.d_gammaR ;
+	gamma = 1.0 / cielab->display.d_gammaR ;
 	cielab->rstep =
 		(cielab->display.d_YCR - cielab->display.d_Y0R)	/ cielab->range;
 	for(i = 0; i <= cielab->range; i++) {
@@ -137,7 +137,7 @@ TIFFCIELabToRGBInit(TIFFCIELabToRGB* cielab,
 	}
 
 	/* Green */
-	gamma = 1.0F / cielab->display.d_gammaG ;
+	gamma = 1.0 / cielab->display.d_gammaG ;
 	cielab->gstep =
 	    (cielab->display.d_YCR - cielab->display.d_Y0R) / cielab->range;
 	for(i = 0; i <= cielab->range; i++) {
@@ -146,7 +146,7 @@ TIFFCIELabToRGBInit(TIFFCIELabToRGB* cielab,
 	}
 
 	/* Blue */
-	gamma = 1.0F / cielab->display.d_gammaB ;
+	gamma = 1.0 / cielab->display.d_gammaB ;
 	cielab->bstep =
 	    (cielab->display.d_YCR - cielab->display.d_Y0R) / cielab->range;
 	for(i = 0; i <= cielab->range; i++) {
